Moves common ui stuff to its own module.
