+++ src/cmd/status.rs
@@ -9,88 +9,7 @@ use git2::{BranchType, Error, Repository, Status, StatusOptions, StatusShow};
 
 use config::{Repo, ReposIterator};
 use invocation::Invocation;
-
-// ----- Severity -------------------------------------------------------------
-
-/// Indicates the severity of a status note.
-#[derive(Clone, PartialEq, PartialOrd)]
-enum Severity {
-    /// Informational (no action needed).
-    Info,
-    /// Local branch is ahead of remote (requires action, but not
-    /// "bad").
-    Notice,
-    /// Uncommitted work or local's being behind or diverging from
-    /// remote (requires action, "bad" situation).
-    Warning,
-}
-
-// ----- Note -----------------------------------------------------------------
-
-/// Represents a piece of information about repository status.
-struct Note {
-    /// Status message that will be displayed to the user.
-    content: String,
-    /// `Severity` of the information.
-    severity: Severity,
-}
-
-impl Note {
-    /// Creates and returns a new `Note`.
-    pub fn new(severity: Severity, content: &str) -> Self {
-        Self {
-            content: content.to_owned(),
-            severity: severity,
-        }
-    }
-
-    /// Returns content of the note.
-    pub fn content(&self) -> &str {
-        &self.content
-    }
-
-    /// Returns severity of the note.
-    pub fn severity(&self) -> &Severity {
-        &self.severity
-    }
-}
-
-// ----- Summary --------------------------------------------------------------
-
-/// Represents the full status of the repository.
-struct Summary {
-    /// `Vec` of `Note` instances representing the status.
-    notes: Vec<Note>,
-}
-
-impl Summary {
-    /// Creates and returns a new `Summary` instance.
-    pub fn new() -> Self {
-        Self { notes: Vec::new() }
-    }
-
-    /// Returns a reference to the `Vec` of `Note`s for this instance.
-    pub fn notes(&self) -> &Vec<Note> {
-        &self.notes
-    }
-
-    /// Adds a new `Note` to this summary.
-    pub fn add_note(&mut self, severity: Severity, content: &str) {
-        self.notes.push(Note::new(severity, content));
-    }
-
-    /// Returns the most severe `Severity` of this summary's notes.
-    pub fn severity(&self) -> Severity {
-        let mut rv = Severity::Info;
-        for note in &self.notes {
-            let s = note.severity();
-            if *s > rv {
-                rv = s.clone();
-            }
-        }
-        rv
-    }
-}
+use ui::{Severity, Summary};
 
 // ----- Worktree -------------------------------------------------------------
